fix _ Stream 대폭 수정 , 상대경로 설정 # 세부사항 1. Stream 클래스를 전부 고침. 현재 패킷이 잘 전달됨을 확인 2. 상대경로를 통해 svn으로 다른 pc에서 작업시 문제없게 만듬
# 내일 할 일
1. 패킷 데이터 전송 실험 후 어떤 패킷들이 필요할지 packet_id 설정
2. db 상대경로 테스트 하기

+++ KaKaoProgram/Src/Stream.cpp
@@ -1,6 +1,6 @@
 #include "Stream.h"
 #define STREAM_READ(type,retval)\
-size_t size = sizeof(type);\
+INT32 size = sizeof(type);\
 if(this->checkReadBound(size)==false){\
 return;\
 }\
